[49] - Controller logic extracted to service
