+++ app/Infrastructure/Exceptions/Handler.php
@@ -2,6 +2,7 @@
 
 namespace App\Infrastructure\Exceptions;
 
+use App\Application\Exceptions\CoinNotFoundException;
 use App\Application\Exceptions\UserNotFoundException;
 use App\Application\Exceptions\WalletNotFoundException;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
@@ -50,7 +51,11 @@ class Handler extends ExceptionHandler
      */
     public function render($request, Throwable $exception): Response
     {
-        if ($exception instanceof UserNotFoundException || $exception instanceof WalletNotFoundException) {
+        if (
+            $exception instanceof UserNotFoundException ||
+            $exception instanceof WalletNotFoundException ||
+            $exception instanceof CoinNotFoundException
+        ) {
             return response()->json([
                 'description' => $exception->getMessage()
             ], $exception->getCode());
