+++ tests/app/Infrastructure/Controller/PostSellCoinControllerTest.php
@@ -41,10 +41,10 @@ class PostSellCoinControllerTest extends TestCase
             ]
         );
 
-        $response->assertStatus(Response::HTTP_NOT_FOUND);
+        $response->assertStatus(Response::HTTP_BAD_REQUEST);
         $response->assertExactJson(
             [
-                'description' => 'A coin with the specified ID was not found'
+                'description' => 'Coin not found'
             ]
         );
     }
@@ -71,10 +71,10 @@ class PostSellCoinControllerTest extends TestCase
             ]
         );
 
-        $response->assertStatus(Response::HTTP_NOT_FOUND);
+        $response->assertStatus(Response::HTTP_BAD_REQUEST);
         $response->assertExactJson(
             [
-                'description' => 'A wallet with the specified ID was not found'
+                'description' => 'Wallet not found'
             ]
         );
     }
